Move new test files into a separate folder
