+++ package.json
@@ -10,7 +10,7 @@
     "backend": "nodemon server/index.js",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint",
-    "test": "jest --coverage --detectOpenHandles"
+    "test": "jest --coverage --runInBand"
   },
   "repository": {
     "type": "git",
